+++ legend-engine-core/legend-engine-core-identity/legend-engine-identity-extensions/src/main/java/org/finos/legend/engine/shared/core/identity/factory/DefaultIdentityFactory.java
@@ -1,4 +1,4 @@
-// Copyright 2023 Goldman Sachs
+// Copyright 2021 Goldman Sachs
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package org.finos.legend.engine.language.pure.dsl.service.execution;
+package org.finos.legend.engine.shared.core.identity.factory;
 
 import org.eclipse.collections.api.factory.Lists;
 import org.eclipse.collections.api.list.ImmutableList;
@@ -20,9 +20,7 @@ import org.eclipse.collections.api.list.MutableList;
 import org.eclipse.collections.impl.utility.LazyIterate;
 import org.finos.legend.engine.shared.core.identity.Credential;
 import org.finos.legend.engine.shared.core.identity.Identity;
-import org.finos.legend.engine.shared.core.identity.credential.AnonymousCredential;
 import org.finos.legend.engine.shared.core.identity.credential.LegendKerberosCredential;
-import org.finos.legend.engine.shared.core.identity.factory.IdentityFactory;
 import org.finos.legend.engine.shared.core.identity.kerberos.SubjectTools;
 import org.finos.legend.server.pac4j.kerberos.KerberosProfile;
 import org.pac4j.core.profile.CommonProfile;
@@ -30,18 +28,20 @@ import org.pac4j.core.profile.CommonProfile;
 import javax.security.auth.Subject;
 import java.security.Principal;
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
-public final class TestIdentityFactory implements IdentityFactory
+public final class DefaultIdentityFactory implements IdentityFactory
 {
-    public static TestIdentityFactory INSTANCE = new TestIdentityFactory();
+    public static DefaultIdentityFactory INSTANCE = new DefaultIdentityFactory();
+    private static final Identity ANONYMOUS_IDENTITY = new Identity("Anonymous");
 
     @Override
     public Identity makeIdentity(Subject subject)
     {
         if (subject == null)
         {
-            return this.getAnonymousIdentity();
+            return this.makeUnknownIdentity();
         }
         Principal principal = getKerberosPrincipalFromSubject(subject);
         if (principal == null)
@@ -60,16 +60,17 @@ public final class TestIdentityFactory implements IdentityFactory
     @Override
     public Identity makeIdentity(MutableList<CommonProfile> profiles)
     {
-        if (profiles == null || profiles.isEmpty())
+        if (profiles == null)
         {
-            return this.getAnonymousIdentity();
+            return this.makeUnknownIdentity();
         }
         Optional<KerberosProfile> kerberosProfileHolder = this.getKerberosProfile(profiles);
         if (kerberosProfileHolder.isPresent())
         {
             return INSTANCE.makeIdentity(kerberosProfileHolder.get().getSubject());
         }
-        return INSTANCE.makeIdentityForTesting(profiles.get(0).getId());
+
+        return INSTANCE.makeUnknownIdentity();
     }
 
     private Optional<KerberosProfile> getKerberosProfile(MutableList<CommonProfile> profiles)
@@ -77,16 +78,20 @@ public final class TestIdentityFactory implements IdentityFactory
         return Optional.ofNullable(LazyIterate.selectInstancesOf(profiles, KerberosProfile.class).getFirst());
     }
 
+    public Identity makeUnknownIdentity()
+    {
+        return new Identity("_UNKNOWN_");
+    }
+
+    public Identity getAnonymousIdentity()
+    {
+        return ANONYMOUS_IDENTITY;
+    }
+
     @Override
     public Identity makeIdentityForTesting(String name)
     {
         return new Identity(name);
-    }
-
-    @Override
-    public Identity getAnonymousIdentity()
-    {
-        return new Identity("Anonymous");
     }
 
     @Override
@@ -100,12 +105,6 @@ public final class TestIdentityFactory implements IdentityFactory
             {
                 LegendKerberosCredential kerberosCredential = (LegendKerberosCredential) credential;
                 profiles.add(new KerberosProfile(kerberosCredential.getSubject(), null));
-            }
-            else if (credential instanceof AnonymousCredential)
-            {
-                CommonProfile profile = new CommonProfile();
-                profile.setId(identity.getName());
-                profiles.add(profile);
             }
         }
         return profiles;
