+++ legend-engine-core/legend-engine-core-identity/legend-engine-identity-extensions/src/main/java/org/finos/legend/engine/shared/core/identity/kerberos/LocalLoginConfiguration.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package org.finos.legend.engine.shared.core.kerberos;
+package org.finos.legend.engine.shared.core.identity.kerberos;
 
 import javax.security.auth.login.AppConfigurationEntry;
 import javax.security.auth.login.Configuration;
