+++ legend-engine-core/legend-engine-core-identity/pom.xml
@@ -21,6 +21,7 @@
 
     <modules>
         <module>legend-engine-identity-core</module>
+        <module>legend-engine-identity-extensions</module>
     </modules>
 
 </project>
