+++ legend-engine-core/legend-engine-core-shared/legend-engine-shared-core/pom.xml
@@ -62,6 +62,10 @@
             <groupId>org.finos.legend.engine</groupId>
             <artifactId>legend-engine-identity-core</artifactId>
         </dependency>
+        <dependency>
+            <groupId>org.finos.legend.engine</groupId>
+            <artifactId>legend-engine-identity-extensions</artifactId>
+        </dependency>
         <!-- ENGINE -->
 
         <!-- JACKSON -->
