--- legend-engine-core/legend-engine-core-shared/legend-engine-shared-core/src/main/java/org/finos/legend/engine/shared/core/identity/credential/KerberosUtils.java
@@ -1,36 +1,0 @@
-// Copyright 2021 Goldman Sachs
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package org.finos.legend.engine.shared.core.identity.credential;
-
-import org.finos.legend.engine.shared.core.identity.Identity;
-
-import javax.security.auth.Subject;
-import java.security.PrivilegedActionException;
-
-public class KerberosUtils
-{
-
-    public static <T> T doAs(final Identity identity, final java.security.PrivilegedAction<T> action)
-    {
-        LegendKerberosCredential wrapper = identity.getCredential(LegendKerberosCredential.class).get();
-        return Subject.doAs(wrapper.getSubject(), action);
-    }
-
-    public static <T> T doAs(final Identity identity, final java.security.PrivilegedExceptionAction<T> action) throws PrivilegedActionException
-    {
-        LegendKerberosCredential wrapper = identity.getCredential(LegendKerberosCredential.class).get();
-        return Subject.doAs(wrapper.getSubject(), action);
-    }
-}
